test: extract so11y test and add logs
