+++ test/e2e/e2e-data/src/main/java/org/apache/skywalking/e2e/metrics/MetricsQuery.java
@@ -115,7 +115,6 @@ public class MetricsQuery extends AbstractQuery<MetricsQuery> {
     public static String[] ALL_SO11Y_LINER_METRICS = {
         METER_INSTANCE_CPU_PERCENTAGE,
         METER_INSTANCE_JVM_MEMORY_BYTES_USED,
-        METER_INSTANCE_TRACE_COUNT,
         METER_INSTANCE_METRICS_FIRST_AGGREGATION,
         METER_INSTANCE_PERSISTENCE_PREPARE_COUNT,
         METER_INSTANCE_PERSISTENCE_EXECUTE_COUNT
