+++ test/e2e/e2e-test/src/test/java/org/apache/skywalking/e2e/simple/SimpleE2E.java
@@ -18,6 +18,8 @@
 
 package org.apache.skywalking.e2e.simple;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.e2e.annotation.ContainerHostAndPort;
 import org.apache.skywalking.e2e.annotation.DockerCompose;
@@ -28,8 +30,6 @@ import org.apache.skywalking.e2e.metrics.AtLeastOneOfMetricsMatcher;
 import org.apache.skywalking.e2e.metrics.Metrics;
 import org.apache.skywalking.e2e.metrics.MetricsQuery;
 import org.apache.skywalking.e2e.metrics.MetricsValueMatcher;
-import org.apache.skywalking.e2e.metrics.ReadMetrics;
-import org.apache.skywalking.e2e.metrics.ReadMetricsQuery;
 import org.apache.skywalking.e2e.retryable.RetryableTest;
 import org.apache.skywalking.e2e.service.Service;
 import org.apache.skywalking.e2e.service.ServicesMatcher;
@@ -56,9 +56,6 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.containers.DockerComposeContainer;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import static org.apache.skywalking.e2e.metrics.MetricsMatcher.verifyMetrics;
 import static org.apache.skywalking.e2e.metrics.MetricsMatcher.verifyPercentileMetrics;
 import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_ENDPOINT_METRICS;
@@ -71,8 +68,6 @@ import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SERVICE_METRICS
 import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SERVICE_MULTIPLE_LINEAR_METRICS;
 import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SERVICE_RELATION_CLIENT_METRICS;
 import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SERVICE_RELATION_SERVER_METRICS;
-import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SO11Y_LABELED_METRICS;
-import static org.apache.skywalking.e2e.metrics.MetricsQuery.ALL_SO11Y_LINER_METRICS;
 import static org.apache.skywalking.e2e.utils.Times.now;
 import static org.apache.skywalking.e2e.utils.Yamls.load;
 
@@ -197,58 +192,6 @@ public class SimpleE2E extends SkyWalkingTestAdapter {
         load("expected/simple/serviceInstanceTopo.yml").as(ServiceInstanceTopologyMatcher.class).verify(topology);
 
         verifyServiceInstanceRelationMetrics(topology.getCalls());
-    }
-
-    @RetryableTest
-    void so11y() throws Exception {
-        List<Service> services = graphql.services(new ServicesQuery().start(startTime).end(now()));
-
-        services = services.stream().filter(s -> s.getLabel().equals("oap::oap-server")).collect(Collectors.toList());
-        LOGGER.info("services: {}", services);
-        load("expected/simple/so11y-services.yml").as(ServicesMatcher.class).verify(services);
-        for (final Service service : services) {
-            final Instances instances = graphql.instances(
-                new InstancesQuery().serviceId(service.getKey()).start(startTime).end(now())
-            );
-
-            LOGGER.info("instances: {}", instances);
-
-            load("expected/simple/so11y-instances.yml").as(InstancesMatcher.class).verify(instances);
-            for (Instance instance : instances.getInstances()) {
-                for (String metricsName : ALL_SO11Y_LINER_METRICS) {
-                    LOGGER.info("verifying service instance response time: {}", instance);
-                    final ReadMetrics instanceMetrics = graphql.readMetrics(
-                        new ReadMetricsQuery().stepByMinute().metricsName(metricsName)
-                            .serviceName(service.getLabel()).instanceName(instance.getLabel())
-                    );
-
-                    LOGGER.info("{}: {}", metricsName, instanceMetrics);
-                    final AtLeastOneOfMetricsMatcher instanceRespTimeMatcher = new AtLeastOneOfMetricsMatcher();
-                    final MetricsValueMatcher greaterThanZero = new MetricsValueMatcher();
-                    greaterThanZero.setValue("gt 0");
-                    instanceRespTimeMatcher.setValue(greaterThanZero);
-                    instanceRespTimeMatcher.verify(instanceMetrics.getValues());
-                }
-                for (String metricsName : ALL_SO11Y_LABELED_METRICS) {
-                    LOGGER.info("verifying service instance response time: {}", instance);
-                    final List<ReadMetrics> instanceMetrics = graphql.readLabeledMetrics(
-                        new ReadMetricsQuery().stepByMinute().metricsName(metricsName)
-                            .serviceName(service.getLabel()).instanceName(instance.getLabel())
-                    );
-
-                    LOGGER.info("{}: {}", metricsName, instanceMetrics);
-                    Metrics allValues = new Metrics();
-                    for (ReadMetrics readMetrics : instanceMetrics) {
-                        allValues.getValues().addAll(readMetrics.getValues().getValues());
-                    }
-                    final AtLeastOneOfMetricsMatcher instanceRespTimeMatcher = new AtLeastOneOfMetricsMatcher();
-                    final MetricsValueMatcher greaterThanZero = new MetricsValueMatcher();
-                    greaterThanZero.setValue("gt 0");
-                    instanceRespTimeMatcher.setValue(greaterThanZero);
-                    instanceRespTimeMatcher.verify(allValues);
-                }
-            }
-        }
     }
 
     private Instances verifyServiceInstances(final Service service) throws Exception {
